+++ src/main/java/org/kdb/inside/brains/lang/refactoring/QRenamePsiElementProcessor.java
@@ -5,7 +5,6 @@ import com.intellij.refactoring.rename.RenamePsiElementProcessor;
 import org.jetbrains.annotations.NotNull;
 import org.kdb.inside.brains.psi.QVariable;
 
-@Deprecated
 public class QRenamePsiElementProcessor extends RenamePsiElementProcessor {
     @Override
     public boolean canProcessElement(@NotNull PsiElement element) {