+++ src/main/java/org/kdb/inside/brains/psi/refs/QImportReferenceProvider.java
@@ -1,13 +1,15 @@
 package org.kdb.inside.brains.psi.refs;
 
 import com.intellij.openapi.util.TextRange;
-import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiReference;
-import com.intellij.psi.PsiReferenceProvider;
+import com.intellij.psi.*;
 import com.intellij.psi.impl.source.resolve.reference.impl.providers.FileReferenceSet;
 import com.intellij.util.ProcessingContext;
 import org.jetbrains.annotations.NotNull;
 import org.kdb.inside.brains.psi.QImport;
+
+import java.util.Collection;
+
+import static java.util.Collections.emptyList;
 
 public class QImportReferenceProvider extends PsiReferenceProvider {
     @Override
@@ -24,10 +26,15 @@ public class QImportReferenceProvider extends PsiReferenceProvider {
 
         final int startOffset = range.getStartOffset();
         final String text = range.substring(element.getText()).trim();
+
         return new FileReferenceSet(text, element, startOffset, null, true, false) {
             @Override
-            public boolean isAbsolutePathReference() {
-                return true;
+            public @NotNull Collection<PsiFileSystemItem> computeDefaultContexts() {
+                final PsiFile file = getContainingFile();
+                if (file == null) {
+                    return emptyList();
+                }
+                return getAbsoluteTopLevelDirLocations(file);
             }
         }.getAllReferences();
     }